everything starts up correctly
Everything starts up correctly but something weird is happening when I
send the request. The logs of the proxy are the following:
```
starting proxy server in port 8080
2024/01/16 01:07:05 starting tunnel
2024/01/16 01:07:09 getting endpoint
2024/01/16 01:07:09 sending request
2024/01/16 01:07:09 200
2024/01/16 01:07:09 service is up
2024/01/16 01:07:09 proxying request to http://127.0.0.1:44379
```
So it seems that the "host" container is able to reach the service
correctly, but when proxying the request it just hangs there and never
returns. The other odd thing is that using the TUI I can't look at the
logs of the `webhook` process, it shows me nothing, not even the startup
logs nor the logs of the request to which responded with a 200 (in the previous
iteration it used to show me all the logs, including the startup ones).
To make it even more strange, after I ctrl-c the dagger up command and
re-run it it never actually starts, it just hangs trying to create a
start a new session.