Add list of files changed in commit to the payload
In order for dispatcher modules to do pattern matching like on github
workflows they need a list of the files that changed. This is a
constraint that all modules will need so it makes sense to add it to
pocketci's agent itself. With this list of files plus some helpers
developers can do the same kind of matching that is done in github
workflows.

There was a small trade off I had to take in order to make this work.
The built-in `Git` function of the dagger SDK does a very weird shallow
clone that contains a few submodules of the dagger repo and basically no
history of the repository being cloned. If we want to know which files
changed we need at least two commits so that we can perform a
comparisson and get the diff. For this reason I had to make a manual git
clone and use `diff-tree` to compute said changes. All of that happens
i
nside dagger containers and we then store that list in the payload that
dispatcher modules receive
Signed-off-by: Matias Pan <matipan@hey.com>