Use BaseRef and Base.SHA for diffing
When in pull requests, github workflows and alike all consider the diff
of the entire pull request, not just each commit that gets pushed. In
order to make sure we behave the same way I added some logic to do a
diff against the base ref when a pull request event happens. The code is
a bit of a mess, needs a bit of love for sure. For PushEvent this change
is not necessary since we already behave correctly (by looking at the
previous commit)

Signed-off-by: Matias Pan <matipan@hey.com>