Deprecate webhook service
This might be rolled back in the future. But since we are going in a
direction of being a tool focused on a developer workflow that is
tightly integrated with the "traditional" way of developing on platforms
like GitHub and GitLab there isn't much value in keeping the webhook
binary around. The "value" that this webhook provided was simply
validating github's header. But that can be done very simply by us. I'm
not sure if we are removing a lot of overhead, but we at least don't
have to deal with webhooks yaml file and we can just `dagger call` our
way to dispatch

Signed-off-by: Matias Pan <matipan@hey.com>