agent: setup a new pocketci agent with support for dagger dispatchers (#6)
* Refactor into agent and server components

Signed-off-by: Matias Pan <matipan@hey.com>

* Add list of files changed in commit to the payload

In order for dispatcher modules to do pattern matching like on github
workflows they need a list of the files that changed. This is a
constraint that all modules will need so it makes sense to add it to
pocketci's agent itself. With this list of files plus some helpers
developers can do the same kind of matching that is done in github
workflows.

There was a small trade off I had to take in order to make this work.
The built-in `Git` function of the dagger SDK does a very weird shallow
clone that contains a few submodules of the dagger repo and basically no
history of the repository being cloned. If we want to know which files
changed we need at least two commits so that we can perform a
comparisson and get the diff. For this reason I had to make a manual git
clone and use `diff-tree` to compute said changes. All of that happens
i
nside dagger containers and we then store that list in the payload that
dispatcher modules receive
Signed-off-by: Matias Pan <matipan@hey.com>

* Allow users to specify top level secrets

Signed-off-by: Matias Pan <matipan@hey.com>

* Use BaseRef and Base.SHA for diffing

When in pull requests, github workflows and alike all consider the diff
of the entire pull request, not just each commit that gets pushed. In
order to make sure we behave the same way I added some logic to do a
diff against the base ref when a pull request event happens. The code is
a bit of a mess, needs a bit of love for sure. For PushEvent this change
is not necessary since we already behave correctly (by looking at the
previous commit)

Signed-off-by: Matias Pan <matipan@hey.com>

* Add list of views that are active

Signed-off-by: Matias Pan <matipan@hey.com>

* Deprecate webhook service

This might be rolled back in the future. But since we are going in a
direction of being a tool focused on a developer workflow that is
tightly integrated with the "traditional" way of developing on platforms
like GitHub and GitLab there isn't much value in keeping the webhook
binary around. The "value" that this webhook provided was simply
validating github's header. But that can be done very simply by us. I'm
not sure if we are removing a lot of overhead, but we at least don't
have to deal with webhooks yaml file and we can just `dagger call` our
way to dispatch

Signed-off-by: Matias Pan <matipan@hey.com>

* Upgrade to latest dagger version

Signed-off-by: Matias Pan <gh@matiaspan.dev>

* Fix & improve Dagger module (#8)

* module: fix go build

Signed-off-by: Solomon Hykes <solomon@dagger.io>

* go mod tidy

Signed-off-by: Solomon Hykes <solomon@dagger.io>

* Make the Dagger module 0.12-ready

This doesn't upgrade the module to 0.12, but makes it so it can be at any time,
with no change of behavior.

Signed-off-by: Solomon Hykes <solomon@dagger.io>

* Fix dagger commands in README

Signed-off-by: Solomon Hykes <solomon@dagger.io>

* module: optimize Dagger API calls

Signed-off-by: Solomon Hykes <solomon@dagger.io>

---------

Signed-off-by: Solomon Hykes <solomon@dagger.io>

* Pocketci builds pocketci

Signed-off-by: Matias Pan <gh@matiaspan.dev>

* Use `eventTrigger` instead of `payload`

Signed-off-by: Matias Pan <gh@matiaspan.dev>

* Update README with up to date instructions

Signed-off-by: Matias Pan <gh@matiaspan.dev>

* Support for commit push event

Signed-off-by: Matias Pan <gh@matiaspan.dev>

* Test cases for supported events

Signed-off-by: Matias Pan <gh@matiaspan.dev>

---------

Signed-off-by: Matias Pan <matipan@hey.com>
Signed-off-by: Matias Pan <gh@matiaspan.dev>
Signed-off-by: Solomon Hykes <solomon@dagger.io>
Co-authored-by: Solomon Hykes <solomon@dagger.io>