Add support for prefix-based multi dispatch (#20)
* Add support for prefix-based multi dispatch

The only way to trigger pocketci today is define a single dispatching function
(`Dispatch` for the lowest of levels, `OnGithub` and so on for higher levels)
in your module. This presents two problems: i) modules with constructors that
have custom parameters are not supported; ii) only one module per repository
is supported. In many cases this is very limiting (e.g in the case of dagger/dagger).
Having the ability to support multiple modules per github event is a must,
mostly for monorepos. On top of that, each module should be able to have more than
one dagger call be triggered from a given event. With this change, functions can
now have a Prefix and pocketci will call every function that ends with the
relevant string instead. This means you can do things such as:
LintGoOnGithubPullRequest and LintElixirOnGithubPullRequest within a single
module, pocketci will call both of those functions upon receiving a github pull
request.

Signed-off-by: Matias Pan <gh@matiaspan.dev>

* Run both Lint and Test as separate dagger calls

Signed-off-by: Matias Pan <gh@matiaspan.dev>

* Lint fixes

Signed-off-by: Matias Pan <gh@matiaspan.dev>

* Add short instructions to the README

Signed-off-by: Matias Pan <gh@matiaspan.dev>

* Fix govet loopcapture

Signed-off-by: Matias Pan <gh@matiaspan.dev>

* Last fixes to the README

Signed-off-by: Matias Pan <gh@matiaspan.dev>

---------

Signed-off-by: Matias Pan <gh@matiaspan.dev>